Reorganized and renamed the whoel project.
